requirements.txt
streamlit
polygon-api-client
google-genai
pandas
numpy
matplotlib
pandas_ta